Use stream and verification.

+++ src/main/java/com/funcxy/oj/controllers/ProblemController.java
@@ -5,11 +5,18 @@ import com.funcxy.oj.repositories.ProblemRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
+
+import static com.funcxy.oj.utils.UserUtil.isSignedIn;
 
 /**
  * Created by wtupc96 on 2017/2/28.
@@ -25,13 +32,17 @@ public class ProblemController {
     MongoTemplate mongoTemplate;
 
     @RequestMapping(method = RequestMethod.POST)
-    public Problem saveProblem(@Valid Problem problem) {
-        return problemRepository.save(problem);
+    public ResponseEntity<Object> saveProblem(@Valid Problem problem, HttpSession session) {
+        if(!isSignedIn(session))
+            return new ResponseEntity<>(new Error(), HttpStatus.FORBIDDEN);
+        return new ResponseEntity<>(problemRepository.save(problem), HttpStatus.OK);
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-    public Problem getOneSpecificProblem(@PathVariable ObjectId id) {
-        return problemRepository.findById(id);
+    public ResponseEntity<Object> getOneSpecificProblem(@PathVariable ObjectId id, HttpSession session) {
+        if(!isSignedIn(session))
+            return new ResponseEntity<>(new Error(), HttpStatus.FORBIDDEN);
+        return new ResponseEntity<>(problemRepository.findById(id), HttpStatus.OK);
     }
 
 //    @RequestMapping(method = RequestMethod.GET)
@@ -53,7 +64,9 @@ public class ProblemController {
 //    }
 
     @RequestMapping(method = RequestMethod.GET)
-    public List<Problem> getProblem(Problem problem) {
+    public ResponseEntity<Object> getProblem(Problem problem, HttpSession session) {
+        if(!isSignedIn(session))
+            return new ResponseEntity<>(new Error(), HttpStatus.FORBIDDEN);
         List<Problem> problemIdList = new ArrayList<>();
         if (problem.getType() != null) {
             problemIdList.addAll(problemRepository.findByTheArg("type", problem.getType()));
@@ -74,23 +87,44 @@ public class ProblemController {
             }
         }
 
-        return problemIdList;
+        class cleanedProblem implements Serializable {
+            public ObjectId id;
+            public String title;
+
+            public cleanedProblem(ObjectId id, String title) {
+                this.id = id;
+                this.title = title;
+            }
+        }
+
+
+        return
+                new ResponseEntity<>
+                        (problemIdList
+                                .stream()
+                                .map(pro ->
+                                        new cleanedProblem(pro.getId(), pro.getTitle()))
+                                .collect(Collectors.toList()), HttpStatus.OK);
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
-    public Problem updateProblem(@RequestBody @Valid Problem problem, @PathVariable ObjectId id) {
+    public ResponseEntity<Object> updateProblem(@RequestBody @Valid Problem problem, @PathVariable ObjectId id, HttpSession session) {
+        if(!isSignedIn(session))
+            return new ResponseEntity<>(new Error(), HttpStatus.FORBIDDEN);
         Problem tempProblem = problemRepository.findById(id);
         if (problem.getReferenceAnswer() != null) {
             problem.setReferenceAnswer(tempProblem.getReferenceAnswer());
         }
         problem.setId(id);
-        return problemRepository.save(problem);
+        return new ResponseEntity<Object>(problemRepository.save(problem), HttpStatus.OK);
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-    public Problem deleteProblem(@PathVariable ObjectId id) {
+    public ResponseEntity<Object> deleteProblem(@PathVariable ObjectId id, HttpSession session) {
+        if(!isSignedIn(session))
+            return new ResponseEntity<>(new Error(), HttpStatus.FORBIDDEN);
         Problem tempProblem = problemRepository.findById(id);
         problemRepository.delete(tempProblem);
-        return tempProblem;
+        return new ResponseEntity<>(tempProblem, HttpStatus.OK);
     }
 }
